Remove blaze-client calls to github.com
Leaves one external SSL call and replaces the other specs with
calls to an internal Jetty server.  Hoping this will reduce the
number of Travis CI timeouts while talking to external services.

This is suggestion #2 from @rossabaker [from Gitter](https://gitter.im/http4s/http4s?at=58a73413aa800ee52cabed70).

Refs #817, #858
